requirements.txt
faiss-cpu
pdfplumber
requests
streamlit
numpy
pandas
dotenv
